requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
httpx==0.25.2
python-dotenv==1.0.0
jinja2==3.1.2
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
typing-extensions==4.8.0
pyyaml==6.0.1
ruamel.yaml==0.18.6
crewai==0.28.0
langchain>=0.1.10,<0.2.0
